src/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
atomicwrites==1.4.0
attrs==20.2.0
black==20.8b1
click==7.1.2
colorama==0.4.3
Django==3.1.1
django-model-utils==4.0.0
djangorestframework==3.11.1
factory-boy==3.0.1
Faker==4.1.3
iniconfig==1.0.1
model-bakery==1.1.1
more-itertools==8.5.0
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pyparsing==2.4.7
pytest==6.0.2
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
regex==2020.7.14
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
